P4-3409 replacing majority of double use (where possible) with BigDecimals for accuracy/precision.

+++ src/main/kotlin/uk/gov/justice/digital/hmpps/pecs/jpc/service/spreadsheet/inbound/price/PricesSpreadsheet.kt
@@ -49,7 +49,7 @@ class PricesSpreadsheet(
     val toLocationName =
       row.getFormattedStringCell(TO_LOCATION) ?: throw RuntimeException("To location name cannot be blank")
 
-    val price = Result.runCatching { Money.valueOf(row.getCell(PRICE).numericCellValue).pence }
+    val price = Result.runCatching { Money.valueOf(row.getCell(PRICE).numericCellValue.toBigDecimal()).pence }
       .onSuccess { if (it == 0) throw RuntimeException("Price must be greater than zero") }
       .getOrElse { throw RuntimeException("Error retrieving price for supplier '$supplier'", it) }
